net: sockets: Allow to register custom poll function
Allow socket engines to register custom `poll` function, if for any
reason it cannot use the native implementation.

Along with the custom `poll` implementation, a socket engine has to
register a helper function, which will check if the fds list provided
refers to sockets owned by that specific engine. If the check is
positive, the custom `poll` is called. Socket engines that are able to
use the native `poll` implementation, should not provide this helper
function, and register a NULL pointer instead.

Note, that once custom `poll` function is registered, it should only be
used with sockets of that specific type. Every custom impelmentation
should verify that it owns the file descriptors provided in the fds
list.

Signed-off-by: Robert Lubos <robert.lubos@nordicsemi.no>